infra/requirements.txt
google-cloud-monitoring==2.15.0
